requirements.txt
async-timeout==4.0.2
beautifulsoup4==4.9.3
blinker==1.6.2
boto3==1.26.165
botocore==1.29.165
certifi==2020.11.8
cffi==1.15.1
chardet==3.0.4
click==8.1.3
dnspython==2.0.0
email-validator==1.1.2
Flask==2.3.0
Flask-Cors==3.0.10
Flask-Login==0.5.0
Flask-WTF==0.14.3
idna==2.10
importlib-metadata==6.6.0
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
lxml==4.6.1
MarkupSafe==2.1.2
pycparser==2.21
pymongo==3.11.0
PyNaCl==1.5.0
python-dateutil==2.8.2
redis==4.5.5
requests==2.25.0
s3transfer==0.6.1
six==1.16.0
soupsieve==2.0.1
urllib3==1.26.2
Werkzeug==2.3.4
WTForms==2.3.3
zipp==3.15.0
